+++ src/common/constants/marcTypes.js
@@ -1,0 +1,4 @@
+export const MARC_TYPES = {
+  BIB: 'bib',
+  HOLDINGS: 'holdings',
+};